feat(utils): hasProperty 추가
