+++ .changeset/short-elephants-kick.md
@@ -1,0 +1,5 @@
+---
+'@modern-kit/utils': minor
+---
+
+feat(utils): hasProperty 추가
