+++ packages/utils/src/validator/hasProperty/index.ts
@@ -1,0 +1,9 @@
+export const hasProperty = <
+  T extends Record<PropertyKey, any>,
+  K extends PropertyKey
+>(
+  obj: T,
+  key: K
+): key is K & keyof T => {
+  return Object.prototype.hasOwnProperty.call(obj, key);
+};
